aiva-backend/requirements.txt
# AIVA Fashion E-commerce Backend Requirements
# Python 3.9+ Required
# Version 2.0 - Italian Fashion with Streaming Support

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# HTTP Client for AI calls
httpx==0.25.2
aiohttp==3.9.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# OpenAI Integration (for production)
openai==1.3.7

# Async Support
aiofiles==23.2.1
asyncio==3.4.3

# Italian Language Support (optional)
unidecode==1.3.7  # For text normalization
langdetect==1.0.9  # For language detection

# JSON handling
orjson==3.9.10

# Logging
colorlog==6.8.0

# Rate limiting
slowapi==0.1.9

# Caching (optional, for production)
redis==5.0.1
aiocache==0.12.2

# Testing (development only)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Performance monitoring (optional)
prometheus-client==0.19.0

# Development tools (optional)
black==23.12.1  # Code formatting
flake8==7.0.0   # Linting
mypy==1.8.0     # Type checking

# Vector/Embedding support (future enhancement)
# Uncomment if implementing vector search
# numpy==1.26.2
# scikit-learn==1.3.2
# sentence-transformers==2.2.2  # For embeddings
# faiss-cpu==1.7.4  # For similarity search